+++ src/container/bodyPanel/routePanels/home/about/About.css
@@ -20,12 +20,6 @@
     padding-top: 15px;
     padding-left: 15px;
 }
-.aboutInfo{
-    font-size: 1vw;
-    font-weight: bold;
-    margin-top: 4%;
-    margin-bottom: 1%;
-}
 .headerLine{
     width: 6em;
     background-color: orange !important;
@@ -35,9 +29,6 @@
     width: 6em;
     background-color: orange !important;
     height: 0.5vh !important;
-}
-.aboutDivider{
-    width: 80%;
 }
 
 @media screen and (max-width: 600px) {
@@ -50,13 +41,6 @@
     }
     .aboutSummary{
         font-size: 4vw;
-    }
-    .aboutInfo{
-        font-size: 2vw;
-        font-weight: bold;
-    }
-    .aboutDivider{
-        width: 85%;
     }    
     .info{
         margin-top: 6px !important;
