+++ src/container/bodyPanel/sideBar/navigation/Navigation.js
@@ -4,7 +4,7 @@ import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
 import Divider from '@material-ui/core/Divider';
-import routes from '../../../../static/routes/routes'
+import routes from '../../../../static/data/routes'
 
 function NavigationItem(props){
     return(
